restucture sections to pass containers implicitly
Added a new feature that allows to pass containers implicitly. Each Section-like
component checks all children and collapses non-Section-like components into
Container and passes 'container'-prop to Section-like component.

Section-like component is the component that contains 'Section' in the name.

'container'-prop is 'true' by default. It can be boolean or React node what
actually is a container.