make switcher-item's links accessible by pointer
Added z-index to fix pointer accessibility of links (mouse clicks
and touches).

There is another accessibility issue: all the the links are accessible
by focus (Tab key) but text is not showing in that moment. Text
and active button must be in sync when user tries to tab over the page.