Fix NetlifyCMS after removing trailing slashes
When there is `trailingSlash: false` in the `next.config.js`,
the framework always redirect to the route without slash. Hence,
the URL under `/public/edit/*` is not resolved.

This changes the NetlifyCMS deployment to resolve the configuration
always from the specific location and moves `/public/edit/index.html`
to bare `/public/edit.html` preventing the confusion with the trailing
slash redirect.

Caused by b79a5a7