Check for the `/statistics` response status (#386)
ElasticSearch is unstable at the moment. API still responds with
a valid JSON object but with unexpected fields. This causes
template tags preserving on the website. It's better if the build
just fails.

This adds checking for the status code in the API response
and improves reporting if there was an attempt to load stats
from the cache in the production environment.

Other behaviours could be considered 🙂