Add HistoryMilestone component
The HistoryMilestone component is a section with image and text
of equal size side by side on desktop and from top to the bottom
on mobile.

The section has a line emulating the timeline and a circle on top of it
emulating history point. It looks good on desktop but might
be questionable on mobile with some pictures.